src/main/java/com/tiketeer/Tiketeer/domain/member/annotation/PasswordValidator.java
package com.tiketeer.Tiketeer.domain.member.annotation;

import java.util.regex.Pattern;

import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;

public class PasswordValidator implements ConstraintValidator<ValidPassword, String> {

	@Override
	public void initialize(ValidPassword constraintAnnotation) {
	}

	private static final String PASSWORD_PATTERN =
		"^(?=.*[a-z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,16}$";

	private final Pattern pattern = Pattern.compile(PASSWORD_PATTERN);

	@Override
	public boolean isValid(String password, ConstraintValidatorContext context) {
		if (password == null) {
			return false;
		}
		return pattern.matcher(password).matches();
	}
}
